requirements.txt
accelerate==1.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.3
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
huggingface-hub==0.26.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.1
pillow==11.0.0
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pytesseract==0.3.13
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
setuptools==75.2.0
sniffio==1.3.1
starlette==0.41.0
sympy==1.13.1
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.46.0
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
